Update login component
Also, renamed the components package
