+++ src/components/content/index.scss
