+++ src/components/content/leftContent/index.scss
