+++ src/components/content/rightContent/index.scss
