+++ src/components/footer/index.scss
