+++ src/components/popup/index.scss
